+++ R/checkdir.R
@@ -12,7 +12,8 @@
 #' @seealso 
 #'  Internaly the function \link[base]{dir.create} is called.
 #'
-#' @return
+#' @return returns TRUE if a new folder is created, FALSE if not
+#' 
 #' @export
 #'
 #' @examples
@@ -25,8 +26,11 @@ checkdir <- function(path, recursive = TRUE, ...) {
   # check if dir exists, if not, create it
   if (!file.exists(path)) {
     dir.create(path = path, recursive = recursive, ...)
+    out <- TRUE
   } else {
     warning(paste0(path, " - already exists"))
+    out <- FALSE
   }
   
+  return(out)
 }